cleanup: discord bot refactor
